Move formatting into `validate-db.ts`
