+++ package.json
@@ -12,10 +12,8 @@
     "test": "jest",
     "storybook": "storybook dev -p 6006",
     "build-storybook": "storybook build",
-    "validate-db": "ts-node --project scripts/tsconfig.json scripts/validate-db.ts && npm run validate-db:format",
-    "validate-db:format": "ts-node --project scripts/tsconfig.json scripts/format-json.ts -- --check",
-    "fix-db": "ts-node --project scripts/tsconfig.json scripts/validate-db.ts -- --fix && npm run fix-db:format",
-    "fix-db:format": "ts-node --project scripts/tsconfig.json scripts/format-json.ts",
+    "validate-db": "ts-node --project scripts/tsconfig.json scripts/validate-db.ts",
+    "fix-db": "ts-node --project scripts/tsconfig.json scripts/validate-db.ts -- --fix",
     "detect-broken": "ts-node --project scripts/tsconfig.json scripts/broken.ts",
     "generate-api": "ts-node --project scripts/tsconfig.json scripts/generate-api.ts",
     "generate-thumbnails": "ts-node --project scripts/tsconfig.json scripts/generate-thumbnails.ts",
