+++ scripts/validate-db.ts
@@ -1,3 +1,5 @@
+import fs from 'fs/promises';
+import path from 'path';
 import { MODEL_PROPS, validateType } from '../src/lib/model-props';
 import { ArchId, ModelId, TagId, UserId } from '../src/lib/schema';
 import { canonicalizeModelId } from '../src/lib/schema-util';
@@ -8,6 +10,24 @@ interface Report {
     message: string;
     fix?: () => Promise<void>;
 }
+
+const getAllFiles = async (dir: string): Promise<string[]> => {
+    const names = await fs.readdir(dir);
+
+    return (
+        await Promise.all(
+            names.map(async (name) => {
+                const filePath = path.join(dir, name);
+                const stat = await fs.stat(filePath);
+                if (stat.isDirectory()) {
+                    return getAllFiles(filePath);
+                } else {
+                    return filePath;
+                }
+            })
+        )
+    ).flat();
+};
 
 const getReports = async (): Promise<Report[]> => {
     const modelData = await fileApi.models.getAll();
@@ -72,6 +92,26 @@ const getReports = async (): Promise<Report[]> => {
             }
         }
     }
+
+    const jsonFiles = (await getAllFiles('data/')).filter((file) => file.endsWith('.json'));
+    await Promise.all(
+        jsonFiles.map(async (file) => {
+            const content = await fs.readFile(file, 'utf-8');
+            const formatted = JSON.stringify(JSON.parse(content), null, 4);
+
+            if (content !== formatted) {
+                errors.push({
+                    message: `File ${file} is not formatted correctly`,
+                    fix: async () => {
+                        // we have to read the file again, because another fix might have changed it
+                        const content = await fs.readFile(file, 'utf-8');
+                        const formatted = JSON.stringify(JSON.parse(content), null, 4);
+                        await fs.writeFile(file, formatted, 'utf-8');
+                    },
+                });
+            }
+        })
+    );
 
     return errors;
 };
